* fixed time sync * fixed unit position sync

+++ src/TA3D_Network.cpp
@@ -193,7 +193,7 @@ void TA3DNetwork::check()
 			units.unit[sync_msg.unit].V.y = 0.0f;
 
 				// Guess where the unit should be now
-			units.unit[sync_msg.unit].Pos = units.unit[sync_msg.unit].Pos + ((units.current_tick - sync_msg.timestamp) * tick_conv) * units.unit[sync_msg.unit].V;
+			units.unit[sync_msg.unit].Pos = units.unit[sync_msg.unit].Pos + ((((int)units.current_tick) - (int)sync_msg.timestamp) * tick_conv) * units.unit[sync_msg.unit].V;
 
 			units.unit[sync_msg.unit].Angle.y = sync_msg.orientation * 360.0f / 65536.0f;
